1st task now working correctly, created Functions.h with all common fucntions for tasks and some other changes

+++ 2 Лаба/2.cpp
@@ -4,14 +4,14 @@
 #include <iomanip>
 #include <fstream>
 
-//float R(float x) {
+//double R(double x) {
 //	int k = 1000;
-//	return round(x * k) / (float)k;
+//	return round(x * k) / (double)k;
 //}
-float R(float x) { return x; }
+double R(double x) { return x; }
 
-float CountInfelicity(std::vector<float> x_de_ure, std::vector<float> x_de_facto) {
-	float sum_x2 = 0, sum_x = 0;
+double CountInfelicity(std::vector<double> x_de_ure, std::vector<double> x_de_facto) {
+	double sum_x2 = 0, sum_x = 0;
 	for (int i = 0; i < x_de_facto.size(); i++) {
 		sum_x2 += x_de_facto[i] * x_de_facto[i];
 		sum_x += (x_de_facto[i] - x_de_ure[i]) * (x_de_facto[i] - x_de_ure[i]);
@@ -19,7 +19,7 @@ float CountInfelicity(std::vector<float> x_de_ure, std::vector<float> x_de_facto
 	return sum_x / sum_x2;
 }
 
-void LogMatrix(std::vector<std::vector<float>> matrix) {
+void LogMatrix(std::vector<std::vector<double>> matrix) {
 	std::ofstream fout("logs.txt", std::ios::app);
 	for (const auto& vec : matrix) {
 		for (auto& el : vec) {
@@ -31,7 +31,7 @@ void LogMatrix(std::vector<std::vector<float>> matrix) {
 	fout.close();
 }
 
-void LogVector(std::vector<float> vector) {
+void LogVector(std::vector<double> vector) {
 	std::ofstream fout("logs.txt", std::ios::app);
 	for (const auto& el : vector) {
 		fout << el << '\t';
@@ -40,8 +40,8 @@ void LogVector(std::vector<float> vector) {
 	fout.close();
 }
 
-std::vector<std::vector<float>> MultMatrix(std::vector<std::vector<float>> a, std::vector<std::vector<float>> b_) {
-	std::vector<std::vector<float>> c(a.size());
+std::vector<std::vector<double>> MultMatrix(std::vector<std::vector<double>> a, std::vector<std::vector<double>> b_) {
+	std::vector<std::vector<double>> c(a.size());
 	for (int i = 0; i < a.size(); i++) {
 		c[i].resize(b_[i].size());
 		for (int j = 0; j < b_[i].size(); j++) {
@@ -53,9 +53,9 @@ std::vector<std::vector<float>> MultMatrix(std::vector<std::vector<float>> a, st
 	return c;
 }
 
-void PrintMatrix(const std::vector<std::vector<float>>& matrix) {
-	for (const std::vector<float>& vec : matrix) {
-		for (const float& x_ : vec) {
+void PrintMatrix(const std::vector<std::vector<double>>& matrix) {
+	for (const std::vector<double>& vec : matrix) {
+		for (const double& x_ : vec) {
 			std::cout << std::setw(7) << std::setprecision(2) << std::left << x_ << " ";
 		}
 		std::cout << std::endl;
@@ -87,7 +87,7 @@ public:
 
 		// заполняем матрицу А
 		int p = 0, sum_1_line = 0;
-		float sum_1_col = 0.0f;
+		double sum_1_col = 0.0f;
 		for (int i = 0; i < matrix_.size(); i++) {
 			for (int j = 0; j < matrix_.size(); j++) {
 				if (i != j) {
@@ -109,11 +109,11 @@ public:
 		// переписываем матрицу А с к = 0
 		matrix_k_0_ = matrix_;
 
-		float sum_1_col_k_0 = (float)sum_1_line;
-		matrix_k_0_[0][0] = -(float)sum_1_line + 1;
+		double sum_1_col_k_0 = (double)sum_1_line;
+		matrix_k_0_[0][0] = -(double)sum_1_line + 1;
 		sum_1_col_k_0 += matrix_k_0_[0][0];
 		
-		matrix_[0][0] = -(float)sum_1_line + pow(10, -2);
+		matrix_[0][0] = -(double)sum_1_line + pow(10, -2);
 		sum_1_col += matrix_[0][0];
 
 		for (int i = 1; i < matrix_.size(); i++) {
@@ -134,8 +134,8 @@ public:
 	}
 
 	// функция умножения матриц
-	std::vector<float> MultMat(std::vector<float> a, std::vector<std::vector<float>> matrix) {
-		std::vector<float> c(matrix.size());
+	std::vector<double> MultMat(std::vector<double> a, std::vector<std::vector<double>> matrix) {
+		std::vector<double> c(matrix.size());
 
 		for (int i = 0; i < matrix.size(); i++) {
 			c[i] = 0;
@@ -147,11 +147,11 @@ public:
 	}
 
 	// функция решения
-	std::pair<std::vector<float>, std::vector<float>> Solve() {
+	std::pair<std::vector<double>, std::vector<double>> Solve() {
 		int line_length = matrix_.size();
 		// заполянем матрицу гаусса
-		std::vector<std::vector<float>> matrix_transform(line_length);
-		std::vector<std::vector<float>> matrix_transform_k_0(line_length);
+		std::vector<std::vector<double>> matrix_transform(line_length);
+		std::vector<std::vector<double>> matrix_transform_k_0(line_length);
 		for (int i = 0; i < line_length; ++i) {
 			matrix_transform[i].resize(line_length + 1);
 			matrix_transform_k_0[i].resize(line_length + 1);
@@ -177,9 +177,9 @@ public:
 			D_k_0_[k][k] = matrix_transform_k_0[k][k];
 
 			for (int i = k + 1; i < line_length; i++) {
-				float t_k_0 = R(matrix_transform_k_0[i][k] / matrix_transform_k_0[k][k]);
+				double t_k_0 = R(matrix_transform_k_0[i][k] / matrix_transform_k_0[k][k]);
 				matrix_transform_k_0[i][k] = t_k_0;
-				float t = R(matrix_transform[i][k] / matrix_transform[k][k]);
+				double t = R(matrix_transform[i][k] / matrix_transform[k][k]);
 				matrix_transform[i][k] = t;
 
 				for (int j = k + 1; j < line_length; ++j) {
@@ -190,7 +190,7 @@ public:
 		}
 
 		// заполняем L из D
-		std::vector<std::vector<float>> Lt(matrix_.size()), Lt_k_0(matrix_k_0_.size());
+		std::vector<std::vector<double>> Lt(matrix_.size()), Lt_k_0(matrix_k_0_.size());
 		L_.resize(matrix_.size());
 		L_k_0_.resize(matrix_.size());
 		for (int i = 0; i < line_length; ++i) {
@@ -235,7 +235,7 @@ public:
 
 		// обратный ход гаусса
 		// решаем уравнение Ly = b
-		std::vector<float> y(L_.size()), y_k_0(L_k_0_.size());
+		std::vector<double> y(L_.size()), y_k_0(L_k_0_.size());
 		for (int i = 0; i < L_.size(); i++) {
 			y[i] = b_[i];
 			y_k_0[i] = b_k_0_[i];
@@ -252,7 +252,7 @@ public:
 		}
 
 		// решаем уравнение Ltx = z
-		std::vector<float> answer(y.size()), answer_k_0(y_k_0.size());
+		std::vector<double> answer(y.size()), answer_k_0(y_k_0.size());
 		for (int i = y.size() - 1; i >= 0; i--) {
 			answer[i] = y[i];
 			answer_k_0[i] = y_k_0[i];
@@ -267,7 +267,7 @@ public:
 		LogVector(answer_k_0);
 		LogMatrix(L_);
 		LogMatrix(L_k_0_);
-		std::vector<float> d, d_k_0;
+		std::vector<double> d, d_k_0;
 		for (int i = 0; i < D_.size(); i++) {
 			d.push_back(D_[i][i]);
 			d_k_0.push_back(D_k_0_[i][i]);
@@ -286,20 +286,20 @@ public:
 		std::ofstream fout("logs.txt", std::ios::app);
 		fout << "k = 2:" << CountInfelicity(x_, answer) << std::endl << "k = 0:" << CountInfelicity(x_, answer_k_0);
 		fout.close();
-		return std::pair<std::vector<float>, std::vector<float>> {answer, answer_k_0};
+		return std::pair<std::vector<double>, std::vector<double>> {answer, answer_k_0};
 	}
 
 private:
-	std::vector<std::vector<float>> matrix_, L_, D_;
-	std::vector<std::vector<float>> matrix_k_0_, L_k_0_, D_k_0_;
-	std::vector<float> b_, x_, b_k_0_;
+	std::vector<std::vector<double>> matrix_, L_, D_;
+	std::vector<std::vector<double>> matrix_k_0_, L_k_0_, D_k_0_;
+	std::vector<double> b_, x_, b_k_0_;
 };
 
 int main() {
 	remove("logs.txt");
 	Matrix m;
 	std::cout << std::endl;
-	std::pair<std::vector<float>, std::vector<float>> pair = m.Solve();
+	std::pair<std::vector<double>, std::vector<double>> pair = m.Solve();
 	std::cout << std::endl << "Vector with k = -2" << std::endl;
 	for (const auto& el : pair.first) {
 		std::cout << std::setw(7) << std::setprecision(7) << el << " ";
