+++ 4 Лаба/4.cpp
@@ -4,13 +4,13 @@
 #include <iomanip>
 #include <fstream>
 
-float R(float x) {
+double R(double x) {
 	int k = 1000;
-	return round(x * k) / (float)k;
+	return round(x * k) / (double)k;
 }
 
-float CountInfelicity(std::vector<float> x_de_ure, std::vector<float> x_de_facto) {
-	float sum_x2 = 0, sum_x = 0;
+double CountInfelicity(std::vector<double> x_de_ure, std::vector<double> x_de_facto) {
+	double sum_x2 = 0, sum_x = 0;
 	for (int i = 0; i < x_de_facto.size(); i++) {
 		sum_x2 += x_de_facto[i] * x_de_facto[i];
 		sum_x += (x_de_facto[i] - x_de_ure[i]) * (x_de_facto[i] - x_de_ure[i]);
@@ -21,7 +21,7 @@ float CountInfelicity(std::vector<float> x_de_ure, std::vector<float> x_de_facto
 	return sum_x / sum_x2;
 }
 
-void LogMatrix(std::vector<std::vector<float>> matrix) {
+void LogMatrix(std::vector<std::vector<double>> matrix) {
 	std::ofstream fout("logs.txt", std::ios::app);
 	for (const auto& vec : matrix) {
 		for (auto& el : vec) {
@@ -33,7 +33,7 @@ void LogMatrix(std::vector<std::vector<float>> matrix) {
 	fout.close();
 }
 
-void LogVector(std::vector<float> vector) {
+void LogVector(std::vector<double> vector) {
 	std::ofstream fout("logs.txt", std::ios::app);
 	for (const auto& el : vector) {
 		fout << el << '\t';
@@ -42,8 +42,8 @@ void LogVector(std::vector<float> vector) {
 	fout.close();
 }
 
-std::vector<std::vector<float>> MultMatrix(std::vector<std::vector<float>> a, std::vector<std::vector<float>> b_) {
-	std::vector<std::vector<float>> c(a.size());
+std::vector<std::vector<double>> MultMatrix(std::vector<std::vector<double>> a, std::vector<std::vector<double>> b_) {
+	std::vector<std::vector<double>> c(a.size());
 	for (int i = 0; i < a.size(); i++) {
 		c[i].resize(b_[i].size());
 		for (int j = 0; j < b_[i].size(); j++) {
@@ -55,7 +55,7 @@ std::vector<std::vector<float>> MultMatrix(std::vector<std::vector<float>> a, st
 	return c;
 }
 
-void PrintMatrix(const std::vector<std::vector<float>>& matrix) {
+void PrintMatrix(const std::vector<std::vector<double>>& matrix) {
 	for (const auto& vec : matrix) {
 		for (const auto& x_ : vec) {
 			std::cout << std::setw(7) << std::setprecision(2) << std::left << x_ << " ";
@@ -64,7 +64,7 @@ void PrintMatrix(const std::vector<std::vector<float>>& matrix) {
 	}
 }
 
-void PrintVector(const std::vector<float>& vector) {
+void PrintVector(const std::vector<double>& vector) {
 	for (const auto& x_ : vector) {
 		std::cout << std::setw(7) << std::setprecision(2) << std::left << x_ << " ";
 	}
@@ -75,8 +75,8 @@ class Matrix {
 public:
 
 	// функция умножения матриц
-	std::vector<float> MultMat(std::vector<float> a, std::vector<std::vector<float>> matrix) {
-		std::vector<float> c(matrix.size());
+	std::vector<double> MultMat(std::vector<double> a, std::vector<std::vector<double>> matrix) {
+		std::vector<double> c(matrix.size());
 
 		for (int i = 0; i < matrix.size(); i++) {
 			c[i] = 0;
@@ -92,7 +92,7 @@ public:
 	}
 
 	// создание матрицы А
-	Matrix(float E, int k_max) : E_(E), k_max_(k_max) {
+	Matrix(double E, int k_max) : E_(E), k_max_(k_max) {
 		srand(time(NULL));
 		// случайное количество элементов матрицы
 		int n = rand() % 3 + 10;
@@ -135,8 +135,8 @@ public:
 		LogMatrix(matrix_);
 	}
 
-	bool EndAlgorithm(std::vector<float> x_new) {
-		float max = 0.0f;
+	bool EndAlgorithm(std::vector<double> x_new) {
+		double max = 0.0f;
 		for (int i = 0; i < x_new.size(); i++) {
 			if (fabs(x_curr_[i] - x_new[i]) > max) {
 				max = fabs(x_curr_[i] - x_new[i]);
@@ -145,7 +145,7 @@ public:
 		return max < E_ || k_ > k_max_;
 	}
 
-	std::vector<float> SolveJacobi() {
+	std::vector<double> SolveJacobi() {
 		// вывод на экран и в файл
 		LogMatrix(matrix_); LogVector(f_);
 		std::cout << "Vector x we need to get: " << std::endl;
@@ -157,10 +157,10 @@ public:
 
 		// метод Якоби
 		x_curr_ = f_;
-		std::vector<float> x_new(matrix_.size());
+		std::vector<double> x_new(matrix_.size());
 		for (k_ = 0; k_ < k_max_; k_++) {
 			for (int i = 0; i < matrix_.size(); i++) {
-				float sum = 0.0f; // считаем суммы для вычисления следующего x
+				double sum = 0.0f; // считаем суммы для вычисления следующего x
 				for (int j = 0; j < matrix_.size(); j++) {
 					if (j != i) {
 						sum += matrix_[i][j] * x_curr_[j];
@@ -186,7 +186,7 @@ public:
 		return x_new;
 	}
 
-	std::vector<float> SolveGaussSeidel(float omega) {
+	std::vector<double> SolveGaussSeidel(double omega) {
 		// вывод на экран и в файл
 		LogMatrix(matrix_); LogVector(f_);
 		std::cout << "Vector x we need to get: " << std::endl;
@@ -198,10 +198,10 @@ public:
 
 		// метод Гаусса-Зейделя
 		x_curr_ = f_;
-		std::vector<float> x_new = x_curr_;
+		std::vector<double> x_new = x_curr_;
 		for (k_ = 0; k_ < k_max_; k_++) {
 			for (int i = 0; i < matrix_.size(); i++) {
-				float sum = 0.0f; // считаем суммы для вычисления следующего x
+				double sum = 0.0f; // считаем суммы для вычисления следующего x
 				for (int j = 0; j < i; j++) {
 					sum += matrix_[i][j] * x_new[j];
 				}
@@ -226,22 +226,22 @@ public:
 	}
 
 private:
-	std::vector<std::vector<float>> matrix_;
-	std::vector<float> f_, x_, x_curr_;
+	std::vector<std::vector<double>> matrix_;
+	std::vector<double> f_, x_, x_curr_;
 	int k_max_, k_;
-	float E_;
+	double E_;
 };
 
 int main() {
 	remove("logs.txt");
 	Matrix m(0.0001, 1000);
 	std::cout << std::endl;
-	std::vector<float> answers = m.SolveJacobi();
+	std::vector<double> answers = m.SolveJacobi();
 	std::cout << "Vector of answers:" << std::endl;
 	PrintVector(answers);
 
 	std::cout << std::endl << std::endl;
-	float omega = 0.5f;
+	double omega = 0.5f;
 	answers = m.SolveGaussSeidel(omega);
 	std::cout << std::endl << "Vector of answers:" << std::endl;
 	PrintVector(answers);
